ADM-361 refactor
