+++ frontend/cypress/fixtures/fixtures.ts
@@ -1,0 +1,4 @@
+export const MOCK_EMAIL = 'test@test.com'
+export const BOARD_TOKEN = 'mockToken'
+export const WEB_SITE = 'https://url.com'
+export const BOARD_PROJECT_KEY = 'mockProjectKey'
